# util/multierr: optimize New for nil cases